Upgrade Django to 1.6.10.
This is a move from Django 1.5.9 to 1.6.10. I have done this
on bedrock witout issue. It may still be a bit harsh to just
do this here, but 1.5.x is unsuported and there are now unpatched
security issues in the wild. This should allow playdoh users at least
a slightly easier path forward.

Not that I've also moved Django from being checked into the repo to
a submodule. We had issues in the past when the official Django github
repo changed the hashes of the previous mirror. Things are now stable
on github for Django, so this seemes easier to update for now.